==> Assay/12.PenalizedLogisticRegression/02.T_wilcox_test.Rmd <==
---
title: "Differential Expression Analysis"
date: "Created: 2021-11-10 Updated: `r Sys.Date()`"
author: 
  - name: "Hua Zou"
    email: "zouhua1@outlook.com"
output: 
  html_notebook:
    codes: hide
---

### Introduction

We perform t-test for protein with normal distribution and wilcoxon-rank-sum test for protein with non-normal distribution on the differential expression analysis.

=====================================================

DataStatus: 

  * Healthy Control: HC
  
  * Adenoma: AA	
  
  * Colorectal Cancer: CRC

* Discovery Cohort: 1112

  * 626 non-CRC samples (481 HC + 145 AA)
  
  * 486 CRC samples

* Validation Cohort: 475 

  * 268 non-CRC samples(206 HC + 62 AA)
  
  * 207 CRC samples
  
=======================================================

Four comparisons between the Colorectal Cancer (CRC) patients and other groups:

  1. CRC 486 vs HC 481
  
  2. CRC 486 vs AA 145
  
  3. AA 145 vs HC 481
  
  4. CRC 486 vs non-CRC( 481 HC; 145 AA) 626

====================================================

```{r setup, include=FALSE}
knitr::opts_chunk$set(warning = F, message = F)
library(dplyr)
library(tibble)
library(convert)
library(ggplot2)

# rm(list = ls())
options(stringsAsFactors = F)
options(future.globals.maxSize = 1000 * 1024^2)

grp <- c("control", "adenoma", "CRC")
subgrp <- c("HC", "AA", "CRC")
grp.col <- c("#568875", "#73FAFC", "#EE853D")
```


### Importing Data
```{r}
ExprSet <- readRDS("../../Result/Biomarker/Logistic_model/species_profile_Discovery.RDS")
table(pData(ExprSet)$Group)
```


### Differential Expression Analysis
```{r}
DEA_test <- function(dataset=ExprSet,
                     group_info="Group",
                     group_name=subgrp[c(3,1)],
                     occurrence=0.2,
                     Pval=0.05,
                     logFC=0.5){
  
  # dataset=ExprSet
  # group_info="Group"
  # group_name=subgrp[c(3,1)]
  # occurrence=0.2
  # Pval=0.05
  # logFC=0.5
  
  
  phenotype <- pData(dataset)
  profile <- exprs(dataset)
  phenotype_cln <- phenotype %>% dplyr::select(all_of(group_info))
  colnames(phenotype_cln)[which(colnames(phenotype_cln) == group_info)] <- "Group"  
  if(length(group_name) == 3){
    group_name2 <- c("CRC", "non_CRC")
    pheno <- phenotype_cln %>%
      rownames_to_column("SampleID") %>%
      filter(Group%in%group_name) %>%
      mutate(Group=ifelse(Group == group_name2[1], group_name2[1], group_name2[2])) %>%
      mutate(Group=factor(Group, levels = group_name2)) %>%
      column_to_rownames("SampleID")
    group_name <- group_name2
  }else{
    pheno <- phenotype_cln %>%
      rownames_to_column("SampleID") %>%
      filter(Group%in%group_name) %>%
      mutate(Group=factor(Group, levels = group_name)) %>%
      column_to_rownames("SampleID")
  }
  print(table(pheno$Group))
  
  edata <- profile[, rownames(pheno)] %>% data.frame() %>% 
      rownames_to_column("tmp") %>%
    filter(apply(dplyr::select(., -one_of("tmp")), 1, function(x) {
            sum(x != 0)/length(x)}) > occurrence) %>%
    column_to_rownames("tmp")
  
  if(length(unique(pheno$Group)) != 2){
    stop("Levels of Group must be 2 levels")
  }
  
  # Checking whether the two groups that where tested were normally distributed using the Shapiro-Wilk test
  shapiro_res <- apply(edata, 1, function(x, y){
    #dat <- data.frame(value=as.numeric(edata[1, ]), group=pheno$Group)
    dat <- data.frame(value=x, group=y)
    t_res <- tapply(dat$value, dat$group, shapiro.test)
    if(t_res[[1]]$p.value < 0.05){
      res <- FALSE
    }else{
      res <- TRUE
    }
    return(res)
  }, pheno$Group) %>%
    data.frame() %>%
    setNames("Normal") %>%
    rownames_to_column("FeatureID")
  
  Normal_prof <- edata[shapiro_res$Normal, ]
  Non_Normal_prof <- edata[!shapiro_res$Normal, ]
  
  if(nrow(Normal_prof) != 0){
    # Normally distributed proteins
    Welch_res <- apply(Normal_prof, 1, function(x, y){
      
      # x=as.numeric(Normal_prof[1, ])
      # y=pheno$Group
  
      dat <- data.frame(value=as.numeric(x), group=y)
      
      # Fold2Change
      dat$value_scale <- scale(dat$value, center = TRUE, scale = TRUE)
      mn_median <- tapply(dat$value_scale, dat$group, median) %>% # compositions::geometricmean
        data.frame() %>% setNames("value") %>%
        rownames_to_column("Group")
      mn_median1 <- with(mn_median, mn_median[Group%in%group_name[1], "value"])
      mn_median2 <- with(mn_median, mn_median[Group%in%group_name[2], "value"])
      if(all(mn_median1 > 0, mn_median2 > 0)){
        Log2FC_median <- log2(mn_median1/mn_median2)
      }else{
        Log2FC_median <- -log2(mn_median1/mn_median2)
      }      
                  
      mn_GM <- tapply(dat$value_scale, dat$group, compositions::geometricmean) %>%
        data.frame() %>% setNames("value") %>%
        rownames_to_column("Group")      
      mn_GM1 <- with(mn_GM, mn_GM[Group%in%group_name[1], "value"])
      mn_GM2 <- with(mn_GM, mn_GM[Group%in%group_name[2], "value"])
      Log2FC_GM <- log2(mn_GM1/mn_GM2)     
      
      # pvalue
      rest <- rstatix::t_test(data = dat, value ~ group)
      
      return(c(Log2FC_median, mn_median1, mn_median2,  
               Log2FC_GM, mn_GM1, mn_GM2, 
               rest$statistic, rest$p))
    }, pheno$Group) %>% 
      t() %>% data.frame() %>%
      setNames(c("logFC_median", paste0("median_", group_name),  
                 "logFC_GM", paste0("geometricmean_", group_name), 
                 "Statistic", "P.value"))
    Normal_res <- Welch_res %>%
      # filter(!is.nan(logFC_GM)) %>%
      # filter(!is.infinite(logFC_GM)) %>%
      rownames_to_column("FeatureID") %>%
      arrange(desc(abs(logFC)), P.value)    
  }else{
    Normal_res <- data.frame()
  }
  
  if(nrow(Non_Normal_prof) != 0){
    # non-Normally distributed proteins
    Wilcox_res <- apply(Non_Normal_prof, 1, function(x, y){
      
      # x=as.numeric(Non_Normal_prof[11, ])
      # y=pheno$Group
      
      dat <- data.frame(value=as.numeric(x), group=y)
      
      # Fold2Change
      dat$value_scale <- scale(dat$value, center = TRUE, scale = TRUE)
      mn_median <- tapply(dat$value_scale, dat$group, median) %>% # compositions::geometricmean
        data.frame() %>% setNames("value") %>%
        rownames_to_column("Group")
      mn_median1 <- with(mn_median, mn_median[Group%in%group_name[1], "value"])
      mn_median2 <- with(mn_median, mn_median[Group%in%group_name[2], "value"])
      if(all(mn_median1 > 0, mn_median2 > 0)){
        Log2FC_median <- log2(mn_median1/mn_median2)
      }else{
        Log2FC_median <- -log2(mn_median1/mn_median2)
      }      
                  
      mn_GM <- tapply(dat$value_scale, dat$group, compositions::geometricmean) %>%
        data.frame() %>% setNames("value") %>%
        rownames_to_column("Group")      
      mn_GM1 <- with(mn_GM, mn_GM[Group%in%group_name[1], "value"])
      mn_GM2 <- with(mn_GM, mn_GM[Group%in%group_name[2], "value"])
      Log2FC_GM <- log2(mn_GM1/mn_GM2) 
      
      # pvalue
      rest <- wilcox.test(data = dat, value ~ group)
      
      return(c(Log2FC_median, mn_median1, mn_median2,  
               Log2FC_GM, mn_GM1, mn_GM2, 
               rest$statistic, rest$p.value))
    }, pheno$Group) %>% 
      t() %>% data.frame() %>%
      setNames(c("logFC_median", paste0("median_", group_name),  
                 "logFC_GM", paste0("geometricmean_", group_name), 
                 "Statistic", "P.value"))
    Non_Normal_res <- Wilcox_res %>%
      # filter(!is.nan(logFC_GM)) %>%
      # filter(!is.infinite(logFC_GM)) %>%
      rownames_to_column("FeatureID") %>%
      arrange(desc(abs(logFC)), P.value)     
  }else{
    Non_Normal_res <- data.frame()
  }
 
  
  # Number & Block
  res <- rbind(Normal_res, Non_Normal_res)
  res$adj.P.Val <- p.adjust(as.numeric(res$P.value), method = "BH")
  
  dat_status <- table(pheno$Group)
  dat_status_number <- as.numeric(dat_status)
  dat_status_name <- names(dat_status)
  res$Block <- paste(paste(dat_status_number[1], dat_status_name[1], sep = "_"),
                         "vs",
                         paste(dat_status_number[2], dat_status_name[2], sep = "_")) 
  # Enrichment Meidan
  res[which(res$logFC_median > logFC & res$adj.P.Val < Pval), "Enrichment_median"] <- group_name[1]
  res[which(res$logFC_median < -logFC & res$adj.P.Val < Pval), "Enrichment_median"] <- group_name[2]
  res[which(abs(res$logFC_median) <= logFC | res$adj.P.Val >= Pval), "Enrichment_median"] <- "Nonsignif"
  
  # Enrichment geometricmean
  res[which(res$logFC_GM > logFC & res$adj.P.Val < Pval), "Enrichment_GM"] <- group_name[1]
  res[which(res$logFC_GM < -logFC & res$adj.P.Val < Pval), "Enrichment_GM"] <- group_name[2]
  res[which(abs(res$logFC_GM) <= logFC | res$adj.P.Val >= Pval), "Enrichment_GM"] <- "Nonsignif"  
  
  res_final <- res %>% dplyr::select(FeatureID, Block, adj.P.Val, P.value, everything()) %>%
    arrange(adj.P.Val) 
  
  return(res_final)
}

if(!dir.exists("../../Result/Biomarker/Logistic_model")){
  dir.create("../../Result/Biomarker/Logistic_model", recursive = T)
}
```


### without Scaling
```{r}
CRC_HC_DEA <- DEA_test(dataset=ExprSet,
                        group_info="Group",
                        group_name=subgrp[c(3,1)],
                        Pval=0.05,
                        logFC=0)

CRC_AA_DEA <- DEA_test(dataset=ExprSet,
                        group_info="Group",
                        group_name=subgrp[c(3,2)],
                        Pval=0.05,
                        logFC=0)

AA_HC_DEA <- DEA_test(dataset=ExprSet,
                      group_info="Group",
                      group_name=subgrp[c(2,1)],
                      Pval=0.05,
                      logFC=0)

CRC_nonCRC_DEA <- DEA_test(dataset=ExprSet,
                             group_info="Group",
                             group_name=subgrp[c(3,2,1)],
                             Pval=0.05,
                             logFC=0)


write.csv(CRC_HC_DEA, "../../Result/Biomarker/Logistic_model/CRC_HC_DEA.csv", row.names = F)
write.csv(CRC_AA_DEA, "../../Result/Biomarker/Logistic_model/CRC_AA_DEA.csv", row.names = F)
write.csv(AA_HC_DEA, "../../Result/Biomarker/Logistic_model/AA_HC_DEA.csv", row.names = F)
write.csv(CRC_nonCRC_DEA, "../../Result/Biomarker/Logistic_model/CRC_nonCRC_DEA.csv", row.names = F)
```


### Final result

Furthermore, the fold change was calculated on a linear scale by dividing the *geometric means* from the two groups. The p values were corrected for multiple testing using the Benjamini–Hochberg method

```{r}
CRC_HC_DEA_final <- CRC_HC_DEA %>% 
  dplyr::select(-all_of(c("logFC_median", "median_CRC", "median_HC", "Enrichment_median"))) %>%
  dplyr::select(FeatureID, Block, adj.P.Val, P.value, logFC_GM, Enrichment_GM, everything()) %>%
  dplyr::rename(logFC=logFC_GM, Enrichment=Enrichment_GM) %>%
  filter(!is.na(Enrichment))
  
CRC_AA_DEA_final <- CRC_AA_DEA %>% 
  dplyr::select(-all_of(c("logFC_median", "median_CRC", "median_AA", "Enrichment_median"))) %>%
  dplyr::select(FeatureID, Block, adj.P.Val, P.value, logFC_GM, Enrichment_GM, everything()) %>%
  dplyr::rename(logFC=logFC_GM, Enrichment=Enrichment_GM) %>%
  filter(!is.na(Enrichment))

AA_HC_DEA_final <- AA_HC_DEA %>% 
  dplyr::select(-all_of(c("logFC_median", "median_AA", "median_HC", "Enrichment_median"))) %>%
  dplyr::select(FeatureID, Block, adj.P.Val, P.value, logFC_GM, Enrichment_GM, everything()) %>%
  dplyr::rename(logFC=logFC_GM, Enrichment=Enrichment_GM) %>%
  filter(!is.na(Enrichment))

CRC_nonCRC_DEA_final <- CRC_nonCRC_DEA %>% 
  dplyr::select(-all_of(c("logFC_median", "median_CRC", "median_non_CRC", "Enrichment_median"))) %>%
  dplyr::select(FeatureID, Block, adj.P.Val, P.value, logFC_GM, Enrichment_GM, everything()) %>%
  dplyr::rename(logFC=logFC_GM, Enrichment=Enrichment_GM) %>%
  filter(!is.na(Enrichment))
table(CRC_nonCRC_DEA_final$Enrichment)

write.csv(CRC_HC_DEA_final, "../../Result/Biomarker/Logistic_model/CRC_HC_DEA_final.csv", row.names = F)
write.csv(CRC_AA_DEA_final, "../../Result/Biomarker/Logistic_model/CRC_AA_DEA_final.csv", row.names = F)
write.csv(AA_HC_DEA_final, "../../Result/Biomarker/Logistic_model/AA_HC_DEA_final.csv", row.names = F)
write.csv(CRC_nonCRC_DEA_final, "../../Result/Biomarker/Logistic_model/CRC_nonCRC_DEA_final.csv", row.names = F) 
```



### systemic information
```{r}
sessionInfo()
```
